backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-ninja==1.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
Pillow==10.1.0
gunicorn==21.2.0
